tests: reorganise integration tests — jubilant at top level, ops_test in subdir
Move jubilant tests to tests/integration/ and ops_test tests to
tests/integration/ops_test/, inverting the previous layout.

- tests/integration/jubilant/{test_dynamic_configs,test_https_multi_unit}.py
  → tests/integration/
- tests/integration/jubilant/conftest.py replaces tests/integration/conftest.py
- tests/integration/test_*.py (ops_test) → tests/integration/ops_test/
- tests/integration/helpers.py → tests/integration/ops_test/helpers.py
- Add tests/integration/ops_test/__init__.py
- Update spread.yaml MODULE paths accordingly
