+++ tests/integration/ops_test/test_basic_auth.py
